requirements.txt
Flask                      # Web framework for creating the UI
python-aiml                # AIML processing library
gunicorn                   # Production WSGI server for deploying Flask apps

